gradient clipping for inception loss

+++ classifier/data_stream.py
@@ -7,7 +7,7 @@ def make_dataset(filenames: List[str],
                  image_size: Tuple[int, int, int]=(224, 224, 3),
                  batch_size: int=32,
                  epochs: int=None) -> tf.data.Iterator:
-    record_defaults = [tf.int32, tf.string]
+    record_defaults = [tf.int64, tf.string]
 
     dataset = tf.contrib.data.CsvDataset(filenames, record_defaults)\
         .map(lambda cid, fp: read_one_image(cid, fp, image_size=image_size)) \
